fixed arm not working in auto
