+++ MeepMeepTesting/build.gradle
@@ -3,15 +3,14 @@ plugins {
 }
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_17
-    targetCompatibility = JavaVersion.VERSION_17
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
 }
 
 repositories {
-    maven { url = 'https://jitpack.io' }
     maven { url = 'https://maven.brott.dev/' }
 }
 
 dependencies {
-    implementation 'com.github.rh-robotics:MeepMeep:v1.0.0'
+    implementation 'com.acmerobotics.roadrunner:MeepMeep:0.1.6'
 }
