+++ settings.gradle
@@ -1,3 +1,4 @@
 include ':FtcRobotController'
 include ':TeamCode'
 include ':MeepMeepTesting'
+include ':MeepMeepTesting'
